// lib/api/generated/model/villeDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Immoplus
 * Immoplus
 * OpenAPI spec version: 1.0
 */

export interface VilleDto {
  id: string;
  name: string;
  createdAt: string;
  createdBy: string;
  updatedAt: string;
  updatedBy: string;
  deletedAt: string;
  deletedBy: string;
}