// lib/api/generated/model/communeDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Immoplus
 * Immoplus
 * OpenAPI spec version: 1.0
 */

export interface CommuneDto {
  id: string;
  name: string;
  ville: string;
  createdAt: string;
  createdBy: string;
  updatedAt: string;
  updatedBy: string;
  deletedAt: string;
  deletedBy: string;
}